Use analysis_name for output filenames
Analysis_name is (currently) the basename of analysis_dir.  Adding
support for running make_dag in a different directory from the current
analysis involved making analysis_dir a full path. Doing that meant that
when I used the analysis_dir (as curdir) as part of the file name of the
output files it was being written to the directory the library.tsv file
was in, instead of the analysis_dir file.