Provide protein coding genes detected plot
This requires having access to the gtf file to find protien coding
genes. Obviously if you don't have a gtf file with protein coding gene
information in it there's no point in showing the plot

Also since I wanted to cache loading the gtf file, I decided to track
how what genomes & annotations were being used for each library, I then
provide a warning if there's more than one combination for an
experiment. As that seems like a weird case, and is probably the result
of a typo.