Break dagman up into smaller pieces
This generates individual dagman scripts for each library and then
splices everything together with a master dagman. Doing this allows
VARS ALL_NODES variables to be local to a specific library which makes
the dagman much, much, shorter. It is really nice to not have to
repeat the paths to software, or the genome description information
over and over.

Additionally the dagman splices gives me an easy way to process a
library and when all the libraries are finished, run the final
reporting.

With the old way there wasn't really a "final" node that could be the
parent for the reporting nodes.