Figure out the flowcell ID on our own for downloads
This looks at the jumpgate web app and pulls down the libraries.
It also merges them into one big fastq in one go by concatinating the
fastqs together. (It also writes a fragments file with the locations
in case you want to reverse the merging)

The one problem is if a library is on multiple flowcells you'll get an
error as it tries to write to the same filename multiple times.

Maybe I should include the flowcell id in the saved filename?