Update pipeline to generate stranded bigwigs.
One change depends on updates to the coverage program over in
GeorgiScripts to read directly from the bam file so the coverage
calculation doesn't need to have the bedgraph files around.

The simpler make_bigwig program can then internally handle the bam to
bedgraph to bigwig calculation. By then using should transfer files,
condor automatically deletes the temporary bedgraph files when it
cleans up the execute directory.

This version is a bit slow as it copies the bam file over before the
calculation, and a bit more cleverness in the condor config might
allow avoiding that.