Implement pyranges based reader.
I had to do some fiddling to the data structure to get it to look like
me earlier methods.

And this version is a inefficient in that it first loads the entire
gtf into the pyranges structure, and then coverts the desired
reference_name chunck into a subselected dataframe to be passed to the
gene/exon region builder.