Replace the genome, annotation, sex fields with a single genome_name
The triplet turned out to be too rigid in practice.

The only place removing it caused problems is with finding the gff/gtf
annotation file. Previously I used the annotation version directly to
compute the annotation file name. This version instead looks for a
file matching the pattern gencode*-tRNAs-ERCC.gff.